Move syntax highlighting and formatting to syles.go and add highlighting to the exampleWrapped and flagUsagesWrapped functions
