+++ cmd/styles.go
@@ -1,6 +1,17 @@
 package cmd
 
-import "github.com/charmbracelet/lipgloss"
+import (
+	"bytes"
+	"fmt"
+	"strings"
+
+	"github.com/alecthomas/chroma/v2/quick"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/muesli/reflow/indent"
+	"github.com/muesli/reflow/wordwrap"
+	"github.com/muesli/reflow/wrap"
+	"github.com/spf13/pflag"
+)
 
 var (
 	width  = 80
@@ -37,4 +48,116 @@ func helpParagraph(para string) string {
 			lipgloss.WithWhitespaceChars("="),
 			lipgloss.WithWhitespaceForeground(subtle),
 		))
+}
+
+// flagUsagesWrapped returns a string containing the usage information
+// for all flags in the FlagSet. Borrowed from pflag, and made ANSI-aware
+// using Reflow. Wrapped to `cols` columns (0 for no wrapping)
+func flagUsagesWrapped(cols int, f *pflag.FlagSet) string {
+	buf := new(bytes.Buffer)
+
+	lines := make([]string, 0, f.NFlag())
+
+	maxlen := 0
+	f.VisitAll(func(flag *pflag.Flag) {
+		if flag.Hidden {
+			return
+		}
+
+		line := ""
+		if flag.Shorthand != "" && flag.ShorthandDeprecated == "" {
+			line = fmt.Sprintf("  -%s, --%s", flag.Shorthand, flag.Name)
+		} else {
+			line = fmt.Sprintf("      --%s", flag.Name)
+		}
+
+		varname, usage := pflag.UnquoteUsage(flag)
+		if varname != "" {
+			line += " " + varname
+		}
+		if flag.NoOptDefVal != "" {
+			switch flag.Value.Type() {
+			case "string":
+				line += fmt.Sprintf("[=\"%s\"]", flag.NoOptDefVal)
+			case "bool":
+				if flag.NoOptDefVal != "true" {
+					line += fmt.Sprintf("[=%s]", flag.NoOptDefVal)
+				}
+			case "count":
+				if flag.NoOptDefVal != "+1" {
+					line += fmt.Sprintf("[=%s]", flag.NoOptDefVal)
+				}
+			default:
+				line += fmt.Sprintf("[=%s]", flag.NoOptDefVal)
+			}
+		}
+
+		// This special character will be replaced with spacing once the
+		// correct alignment is calculated
+		line += "\x00"
+		if len(line) > maxlen {
+			maxlen = len(line)
+		}
+
+		usages := strings.SplitAfterN(usage, "Query:", 2)
+		line += usages[0]
+		if len(usages) > 1 {
+			line += jsonHighlight(usages[1])
+		}
+
+		// if !flag.defaultIsZeroValue() {
+		// 	if flag.Value.Type() == "string" {
+		// 		line += fmt.Sprintf(" (default %q)", flag.DefValue)
+		// 	} else {
+		// 		line += fmt.Sprintf(" (default %s)", flag.DefValue)
+		// 	}
+		// }
+		if len(flag.Deprecated) != 0 {
+			line += fmt.Sprintf(" (DEPRECATED: %s)", flag.Deprecated)
+		}
+
+		lines = append(lines, line)
+	})
+
+	for _, line := range lines {
+		sidx := strings.Index(line, "\x00")
+		spacing := strings.Repeat(" ", maxlen-sidx)
+		fmt.Fprint(buf, line[:sidx])
+		// maxlen + 2 comes from + 1 for the \x00 and + 1 for the (deliberate) off-by-one in maxlen-sidx
+		// try to wrap at cols-8 and if that fails enforce at cols wide
+		wrapped := wrap.String(wordwrap.String(line[sidx+1:], cols-maxlen-10), cols-maxlen-2)
+		for i, subline := range strings.Split(wrapped, "\n") {
+			if i == 0 {
+				fmt.Fprintln(buf, spacing, subline)
+			} else {
+				fmt.Fprintln(buf, indent.String(subline, uint(maxlen+1)))
+			}
+		}
+	}
+
+	return buf.String()
+}
+
+// exampleWrapped indents and wraps the `example` (query) text
+// for tq / cobra commands using ANSI-aware wrapping to `cols` width and a
+// 2-column indent
+func exampleWrapped(cols int, example string) string {
+	buf := new(bytes.Buffer)
+	example = jsonHighlight(example)
+	// try to wrap at cols-8 and if that fails enforce at cols wide
+	wrapped := wrap.String(wordwrap.String(example, cols-8), cols)
+	for _, subline := range strings.Split(wrapped, "\n") {
+		fmt.Fprintln(buf, indent.String(subline, 2))
+	}
+	return buf.String()
+}
+
+// Syntax highlighting for json strings using chroma
+func jsonHighlight(json string) string {
+	w := new(bytes.Buffer)
+	err := quick.Highlight(w, json, "json", "terminal16m", "github")
+	if err != nil {
+		panic(err)
+	}
+	return w.String()
 }
